아이돌 관리, 포토카드 관리 기능에서 작성 폼에 작은따옴표가 있는 경우 발생하는 쿼리 문제 해결

+++ backend/src/routes/photocardCtrl.js
@@ -5,7 +5,7 @@ const fsAsync = require('fs');
 const { db } = require('../config/database');
 const { getTimestampFilename, photocardImageUpload, PHOTOCARD_IMAGE_DIR } = require('../config/multer');
 const { isAdmin, verifyLogin } = require('../utils/jwt');
-const { isNull, getWhereClause } = require('../utils/common');
+const { isNull, getWhereClause, convertToMysqlStr } = require('../utils/common');
 
 // 포토카드 목록 반환
 router.get('/photocard/list', async (req, res) => {
@@ -87,7 +87,7 @@ router.get('/photocard/detail/:photocardId', async (req, res) => {
 
 // 포토카드 등록 처리
 router.post('/photocard', photocardImageUpload.single('image'), verifyLogin, async (req, res) => {
-  const { groupId, memberId, albumId, name } = req.body;
+  let { groupId, memberId, albumId, name } = req.body;
   const { accessToken, file } = req;
 
   // 유효성 검사 실패시 다운 받은 임시 이미지 파일을 삭제하는 함수
@@ -124,6 +124,9 @@ router.post('/photocard', photocardImageUpload.single('image'), verifyLogin, asy
   
   const con = await db.getConnection();
   try {
+    // 쿼리에 사용할 수 있는 형태로 변환
+    name = convertToMysqlStr(name);
+
     // 해당 그룹 존재 여부 확인
     let sql = `SELECT name FROM GroupData WHERE group_id='${groupId}'`;
     let [[group]] = await con.query(sql);
@@ -149,8 +152,10 @@ router.post('/photocard', photocardImageUpload.single('image'), verifyLogin, asy
     }
 
     // DB에 저장
-    sql = `INSERT INTO Photocard (name, group_id, member_id, album_id) VALUES (?, ?, ?, ?)`;
-    let [result] = await con.execute(sql, [name, groupId, memberId, albumId]);
+    sql = `
+    INSERT INTO Photocard (name, group_id, member_id, album_id)
+    VALUES ('${name}', ${groupId}, ${memberId}, ${albumId})`;
+    let [result] = await con.execute(sql);
 
     // 임시로 받은 이미지 파일의 이름을 실제로 저장할 이름으로 변경
     let filename = "";
@@ -177,7 +182,7 @@ router.post('/photocard', photocardImageUpload.single('image'), verifyLogin, asy
 // 포토카드 수정 처리
 router.put('/photocard/:photocardId', photocardImageUpload.single('image'), verifyLogin, async (req, res) => {
   const { photocardId } = req.params;
-  const { groupId, memberId, albumId, name } = req.body;
+  let { groupId, memberId, albumId, name } = req.body;
   const { accessToken, file } = req;
 
   // 유효성 검사 실패시 다운 받은 임시 이미지 파일을 삭제하는 함수
@@ -214,6 +219,9 @@ router.put('/photocard/:photocardId', photocardImageUpload.single('image'), veri
 
   const con = await db.getConnection();
   try {
+    // 쿼리에 사용할 수 있는 형태로 변환
+    name = convertToMysqlStr(name);
+
     // 수정하려는 멤버 존재 유무 확인
     let sql = `SELECT photocard_id, image_name FROM Photocard WHERE photocard_id=${photocardId}`;
     let [[photocard]] = await con.query(sql);
@@ -222,14 +230,14 @@ router.put('/photocard/:photocardId', photocardImageUpload.single('image'), veri
       return res.status(404).json({ message: '수정하려는 포토카드가 DB에 없습니다.' });
     }
 
-    console.log(file);
     // 임시로 받은 이미지 파일의 이름을 실제로 저장할 이름으로 변경하고 기존의 이미지 삭제
     let filename = "";
     if (file) {
-
       // 기존 이미지 삭제
-      try { fs.rm(path.join(file.destination, photocard.image_name)); }
-      catch (err) { console.error(err); }
+      if (photocard.image_name) {
+        try { fs.rm(path.join(file.destination, photocard.image_name)); }
+        catch (err) { console.error(err); }
+      }
 
       // 이미지 이름 변경
       filename = getTimestampFilename(photocardId, file.mimetype);
@@ -282,8 +290,10 @@ router.delete('/photocard/:photocardId', verifyLogin, async (req, res) => {
     await con.execute(sql);
 
     // 이미지 파일 삭제
-    try { fs.rm(path.join(PHOTOCARD_IMAGE_DIR, photocard.image_name)); }
-    catch (err) { console.error(err); }
+    if (photocard.image_name) {
+      try { fs.rm(path.join(PHOTOCARD_IMAGE_DIR, photocard.image_name)); }
+      catch (err) { console.error(err); }
+    }
 
     return res.status(200).json({ message: '포토카드 정보를 삭제했습니다.' });
   } catch (err) {
